+++ .pre-commit-config.yaml
@@ -12,7 +12,7 @@ repos:
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-yaml
-      - id: check-added-large-files
+#      - id: check-added-large-files
       - id: check-toml
       - id: debug-statements
