+++ config.yaml
@@ -36,7 +36,7 @@ jobs:
       ref: Dune
       table_name: dune_sync_test_table
 
-  - name: Sync Parameterized Dune Query with Multiple Types to Postgres
+  - name: Sync Non-parameterized Dune Query with Multiple Types to Postgres
     source:
       ref: Dune
       query_id: 4273244
